Improve UX for user filters for time entries and reports. Rename latestUserFilter to userFilter. Fix layout-shifts.
