+++ js/main.js
@@ -14,6 +14,7 @@ import Timerange from "./views/Timerange.svelte";
 import PrintButton from "./views/PrintButton.svelte";
 import Import from "./views/Import.svelte";
 import UserFilter from "./views/UserFilter.svelte";
+import UserFilterButton from "./views/UserFilterButton.svelte";
 // import Settings from "./views/Settings.svelte";
 import { Helpers } from "./lib/helpers";
 import { PagePjax } from "./lib/pjax";
@@ -227,6 +228,17 @@ const init = () => {
 		safelyCreateComponent({
 			component: Import,
 			selector: "#content.app-timemanager [data-svelte='Import.svelte']",
+			props: {
+				...store,
+				requestToken: token,
+			},
+		})
+	);
+
+	components.push(
+		safelyCreateComponent({
+			component: UserFilterButton,
+			selector: "#content.app-timemanager [data-svelte='UserFilterButton.svelte']",
 			props: {
 				...store,
 				requestToken: token,
