+++ l10n/de_AT.json
@@ -153,7 +153,8 @@
 		"Close": "Schließen",
 		"All projects": "Alle Projekte",
 		"All tasks": "Alle Aufgaben",
-		"Show only entries created by": "Zeige nur Einträge von:"
+		"Created by": "Einträge von",
+		"Filter": "Filter"
 	},
 	"pluralForm": "nplurals=2; plural=(n != 1);"
 }
